+++ p5-server/src/server/directory-listing.ts
@@ -9,17 +9,17 @@ import { templateDir } from './globals';
 const directoryListingTmpl = pug.compileFile(path.join(templateDir, 'directory.pug'));
 
 export async function createDirectoryListing(dir: string, breadcrumbPath?: string): Promise<string> {
-  const { sketches, unaffiliatedFiles } = await Sketch.analyzeDirectory(dir);
-  unaffiliatedFiles.sort((a, b) => a.toLowerCase().localeCompare(b.toLowerCase()));
+  const { sketches, unassociatedFiles } = await Sketch.analyzeDirectory(dir);
+  unassociatedFiles.sort((a, b) => a.toLowerCase().localeCompare(b.toLowerCase()));
 
-  const readmeName = unaffiliatedFiles.find(s => s.toLowerCase() === 'readme.md');
+  const readmeName = unassociatedFiles.find(s => s.toLowerCase() === 'readme.md');
   const readme = readmeName && {
     name: readmeName,
     html: marked(fs.readFileSync(path.join(dir, readmeName), 'utf-8'))
   };
 
-  const directories = unaffiliatedFiles.filter(s => fs.statSync(path.join(dir, s)).isDirectory());
-  const files = unaffiliatedFiles.filter(s => !directories.includes(s) && s !== readmeName);
+  const directories = unassociatedFiles.filter(s => fs.statSync(path.join(dir, s)).isDirectory());
+  const files = unassociatedFiles.filter(s => !directories.includes(s) && s !== readmeName);
   const title = dir === './' ? 'P5 Server' : path.basename(dir);
 
   const pathComponents = pathComponentsForBreadcrumbs(breadcrumbPath || dir);